Factor out common code from RSpace and ReplayRSpace
