+++ rspace/src/main/scala/coop/rchain/rspace/ReplayRSpace.scala
@@ -17,26 +17,21 @@ import scala.collection.immutable.Seq
 import scala.concurrent.SyncVar
 import scala.util.Random
 
-class ReplayRSpace[C, P, A, R, K](val store: IStore[C, P, A, K], val branch: Branch)(
+class ReplayRSpace[C, P, A, R, K](store: IStore[C, P, A, K], branch: Branch)(
     implicit
     serializeC: Serialize[C],
     serializeP: Serialize[P],
     serializeA: Serialize[A],
     serializeK: Serialize[K]
-) extends ISpace[C, P, A, R, K] {
+) extends RSpaceOps[C, P, A, R, K](store, branch)
+    with ISpace[C, P, A, R, K] {
 
-  private val logger: Logger = Logger[this.type]
+  override protected[this] val logger: Logger = Logger[this.type]
 
   private val replayData: SyncVar[ReplayData] = {
     val sv = new SyncVar[ReplayData]()
     sv.put(ReplayData.empty)
     sv
-  }
-
-  private[this] val installs: SyncVar[Installs[C, K, R, A, P]] = {
-    val installs = new SyncVar[Installs[C, K, R, A, P]]()
-    installs.put(Map.empty)
-    installs
   }
 
   def consume(channels: Seq[C], patterns: Seq[P], continuation: K, persist: Boolean)(
@@ -151,59 +146,6 @@ class ReplayRSpace[C, P, A, R, K](val store: IStore[C, P, A, K], val branch: Bra
             }
         }
       }
-    }
-
-  private[this] def install(txn: store.T, channels: Seq[C], patterns: Seq[P], continuation: K)(
-      implicit m: Match[P, A, R]): Option[(K, Seq[R])] = {
-    if (channels.length =!= patterns.length) {
-      val msg = "channels.length must equal patterns.length"
-      logger.error(msg)
-      throw new IllegalArgumentException(msg)
-    }
-    logger.debug(s"""|install: searching for data matching <patterns: $patterns>
-                     |at <channels: $channels>""".stripMargin.replace('\n', ' '))
-
-    val consumeRef = Consume.create(channels, patterns, continuation, true)
-    installs.update(_.updated(channels, Install(patterns, continuation, m)))
-
-    /*
-     * Here, we create a cache of the data at each channel as `channelToIndexedData`
-     * which is used for finding matches.  When a speculative match is found, we can
-     * remove the matching datum from the remaining data candidates in the cache.
-     *
-     * Put another way, this allows us to speculatively remove matching data without
-     * affecting the actual store contents.
-     */
-
-    val channelToIndexedData = channels.map { (c: C) =>
-      c -> Random.shuffle(store.getData(txn, Seq(c)).zipWithIndex)
-    }.toMap
-
-    val options: Option[Seq[DataCandidate[C, R]]] =
-      extractDataCandidates(channels.zip(patterns), channelToIndexedData, Nil).sequence
-
-    options match {
-      case None =>
-        store.removeAll(txn, channels)
-        store.installWaitingContinuation(
-          txn,
-          channels,
-          WaitingContinuation(patterns, continuation, persist = true, consumeRef))
-        for (channel <- channels) store.addJoin(txn, channel, channels)
-        logger.debug(s"""|consume: no data found,
-                         |storing <(patterns, continuation): ($patterns, $continuation)>
-                         |at <channels: $channels>""".stripMargin.replace('\n', ' '))
-        None
-      case Some(_) =>
-        throw new RuntimeException("Installing can be done only on startup")
-    }
-
-  }
-
-  def install(channels: Seq[C], patterns: Seq[P], continuation: K)(
-      implicit m: Match[P, A, R]): Option[(K, Seq[R])] =
-    store.withTxn(store.createTxnWrite()) { txn =>
-      install(txn, channels, patterns, continuation)
     }
 
   def produce(channel: C, data: A, persist: Boolean)(
@@ -373,12 +315,6 @@ class ReplayRSpace[C, P, A, R, K](val store: IStore[C, P, A, K], val branch: Bra
     // update the replay data
     replayData.update(const(rigs))
   }
-
-  override protected[this] def restoreInstalls(txn: store.T): Unit =
-    installs.get.foreach {
-      case (channels, Install(patterns, continuation, _match)) =>
-        install(txn, channels, patterns, continuation)(_match)
-    }
 
 }
